requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
cachetools==7.0.3
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
clerk-backend-api==5.0.2
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
aio-pika==9.6.2
aiormq==6.9.4
packaging==26.0
pika==1.3.2
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.43.0
uvicorn-worker==0.4.0
vine==5.1.0
wcwidth==0.6.0
